+++ examples/CaffeModels/load-cpm.py
@@ -97,7 +97,7 @@ def CPM(image):
 def run_test(model_path, img_file):
     param_dict = dict(np.load(model_path))
     predict_func = OfflinePredictor(PredictConfig(
-        inputs_desc=[InputDesc(tf.float32, (None, 368, 368, 3), 'input')],
+        input_signature=[tf.TensorSpec((None, 368, 368, 3), tf.float32, 'input')],
         tower_func=CPM,
         session_init=DictRestore(param_dict),
         input_names=['input'],