+++ examples/basics/cifar-convnet.py
@@ -64,7 +64,7 @@ class Model(ModelDesc):
         cost = tf.nn.sparse_softmax_cross_entropy_with_logits(logits=logits, labels=label)
         cost = tf.reduce_mean(cost, name='cross_entropy_loss')
 
-        correct = tf.cast(tf.nn.in_top_k(logits, label, 1), tf.float32, name='correct')
+        correct = tf.cast(tf.nn.in_top_k(predictions=logits, targets=label, k=1), tf.float32, name='correct')
         # monitor training error
         add_moving_summary(tf.reduce_mean(correct, name='accuracy'))
 
@@ -76,7 +76,7 @@ class Model(ModelDesc):
         return tf.add_n([cost, wd_cost], name='cost')
 
     def optimizer(self):
-        lr = tf.get_variable('learning_rate', initializer=1e-2, trainable=False)
+        lr = tf.Variable(1e-2, name='learning_rate', trainable=False)
         tf.summary.scalar('lr', lr)
         return tf.train.AdamOptimizer(lr, epsilon=1e-3)