+++ tensorpack/train/interface.py
@@ -87,7 +87,7 @@ def launch_train_with_config(config, trainer):
     # We should gradually stay away from this unuseful abstraction.
     # TowerFuncWrapper is a better abstraction (similar to tf.defun in the future)
     trainer.setup_graph(
-        model.get_inputs_desc(), input,
+        model.get_input_signature(), input,
         model._build_graph_get_cost, model.get_optimizer)
     _check_unused_regularization()
     trainer.train_with_defaults(